+++ src/db_migrate.cpp
@@ -13,7 +13,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.7.2"};
+constexpr auto version = std::string_view{"0.8.0"};
 
 auto print_help() -> void
 {